+++ lib/src/impl/console_quill/color_style/cq_std_styles.dart
@@ -1,14 +1,13 @@
-import '../console_quill.dart';
 import 'cq_style.dart';
 
-/// Style of [ConsoleQuill]'s output.
+/// Standard styles.
 class CQSTDStyles {
   const CQSTDStyles._();
 
   static const normal = CQStyle(0);
 
   static const boldOn = CQStyle(1);
-  static const boldOff = CQStyle(21);
+  static const doublyUnderlinedBoldOff = CQStyle(21);
 
   static const italicOn = CQStyle(3);
   static const italicOff = CQStyle(23);
@@ -16,6 +15,6 @@ class CQSTDStyles {
   static const underlineOn = CQStyle(4);
   static const underlineOff = CQStyle(24);
 
-  static const crossedOn = CQStyle(9);
-  static const crossedOff = CQStyle(29);
+  static const strikeOn = CQStyle(9);
+  static const strikeOff = CQStyle(29);
 }